refacto(services): extract auth function to authApi.js #30
